Success!! Git em the title!!!
I have finally figured how to dynamically get the title of the playlists. I was overthinking it by a longshot while all i had to do was to user queryselector and innerHTML method to do so... Thats coding for ya

Next up is getting all the playlists using the next page token and going from there.