I am finished... for now.
I am technically finished with this project as I sought out what I was trying to do with this project. Most of the time was trying to figure out the auth2.0 process!!!

I am content with how far this project went but if I want to go further with it would mean another long period of time and right now I want to focus on more websites!

This is the end of this chapter, onto the next!